rearranged devices a bit
